Remove PomVersion.getRmbVersion(), move RmbVersion to domain-models-runtime
PomVersion.getRmbVersion() is used by domain-models-runtime only:
https://github.com/search?q=org%3Afolio-org+getrmbversion&type=Code
This change does not affect any RMB based module.
