+++ domain-models-interface-extensions/pom.xml
@@ -107,47 +107,6 @@
     <plugins>
 
       <plugin>
-        <artifactId>maven-resources-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>generate-rmbversion-source</id>
-            <phase>process-resources</phase>
-            <goals>
-              <goal>resources</goal>
-            </goals>
-            <configuration>
-              <outputDirectory>${project.build.directory}/generated-sources/rmbversion</outputDirectory>
-              <resources>
-                <resource>
-                  <directory>src/main/java-filtered</directory>
-                  <filtering>true</filtering>
-                </resource>
-              </resources>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>build-helper-maven-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>add-rmbversion-source</id>
-            <goals>
-              <goal>add-source</goal>
-            </goals>
-            <phase>process-sources</phase>
-            <configuration>
-              <sources>
-                <source>${project.build.directory}/generated-sources/rmbversion</source>
-              </sources>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-
-      <plugin>
         <groupId>com.nickwongdev</groupId>
         <artifactId>aspectj-maven-plugin</artifactId>
         <version>1.12.6</version>
@@ -222,11 +181,6 @@
             </configuration>
           </execution>
         </executions>
-      </plugin>
-
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-failsafe-plugin</artifactId>
       </plugin>
 
     </plugins>
